code/src/smart-email-orchestrator/requirements.txt
Flask==2.0.2
transformers==4.11.3
Pillow==8.4.0
pytesseract==0.3.8
werkzeug==2.0.2
pandas==1.3.4
numpy==1.21.2